requirements.txt
discord.py
requests
python-dotenv
apscheduler
beautifulsoup4
pytz